Squashed commit with OpenNI2Driver fixes and improvements

+++ src/modules/openni2/OpenNI2DeviceDriverClient.h
@@ -55,8 +55,8 @@ public:
     virtual void onRead(ImageOf<PixelRgb>& img);
     virtual void onRead(ImageOf<PixelMono16>& img);
     // IKinectDeviceDriver
-    virtual bool getSkeletonOrientation(Vector *vectorArray, double *confidence,  int userID);
-    virtual bool getSkeletonPosition(Vector *vectorArray, double *confidence,  int userID);
+    virtual bool getSkeletonOrientation(Vector *vectorArray, float *confidence,  int userID);
+    virtual bool getSkeletonPosition(Vector *vectorArray, float *confidence,  int userID);
     virtual int *getSkeletonState();
     virtual int getSkeletonState(int userID);
     virtual ImageOf<PixelRgb> getImageFrame();
